requirements.txt
langchain-openai
langchain
python-dotenv
langchain_community
langchain_core
streamlit
langchain-groq
arxiv
wikipedia
duckduckgo-search
serpapi
SQLAlchemy
mysql-connector-python
validators==0.28.1
youtube_transcript_api
unstructured
numexpr
langchain_huggingface
huggingface_hub 
sentence_transformers
gradio
langchain
